Remove mrjob.conversions and move its contents to mrjob.parse.
