+++ mrjob/parse.py
@@ -13,9 +13,11 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 """Utilities for parsing errors, counters, and status messages."""
+from datetime import datetime
 from functools import wraps
 import logging
 import re
+import time
 from urlparse import ParseResult
 from urlparse import urlparse as urlparse_buggy
 
@@ -27,6 +29,12 @@ except ImportError:
 
 from mrjob.compat import uses_020_counters
 
+try:
+    import boto.utils
+except ImportError:
+    # don't require boto; MRJobs don't actually need it when running
+    # inside hadoop streaming
+    boto = None
 
 # match the filename of a hadoop streaming jar
 HADOOP_STREAMING_JAR_RE = re.compile(r'^hadoop.*streaming.*\.jar$')
@@ -527,3 +535,29 @@ def parse_hadoop_counters_from_line(line, hadoop_version=None):
         counters[group].setdefault(counter, 0)
         counters[group][counter] += int(value)
     return counters, int(m.group('step_num'))
+
+### AWS Date-time parsing ###
+
+# sometimes AWS gives us seconds as a decimal, which we can't parse
+# with boto.utils.ISO8601
+SUBSECOND_RE = re.compile('\.[0-9]+')
+
+
+# Thu, 29 Mar 2012 04:55:44 GMT
+RFC1123 = '%a, %d %b %Y %H:%M:%S %Z'
+
+
+def iso8601_to_timestamp(iso8601_time):
+    iso8601_time = SUBSECOND_RE.sub('', iso8601_time)
+    try:
+        return time.mktime(time.strptime(iso8601_time, boto.utils.ISO8601))
+    except ValueError:
+        return time.mktime(time.strptime(iso8601_time, RFC1123))
+
+
+def iso8601_to_datetime(iso8601_time):
+    iso8601_time = SUBSECOND_RE.sub('', iso8601_time)
+    try:
+        return datetime.strptime(iso8601_time, boto.utils.ISO8601)
+    except ValueError:
+        return datetime.strptime(iso8601_time, RFC1123)
