+++ tests/mockboto.py
@@ -36,9 +36,9 @@ except ImportError:
     boto = None
 
 from mrjob.conf import combine_values
-from mrjob.conversions import RFC1123
 from mrjob.parse import is_s3_uri
 from mrjob.parse import parse_s3_uri
+from mrjob.parse import RFC1123
 
 
 DEFAULT_MAX_JOB_FLOWS_RETURNED = 500
